Storage classes can have prefixes
Adds a `prefix` kwarg to `CloudFilesStorage`, which specifies a string
to be prefixed to the filenames saved with the instance. Also adds
`PREFIX` and `STATIC_PREFIX` settings, which are the default prefix
values for `CloudFilesStorage` and `CloudFilesStaticStorage` instances,
respectively.

This fixes `CloudFilesStaticStorage.save` and
`CloudFilesStaticStorage.url`, which use unprefixed names, making them
incompatible with the syncstatic command (which always used `STATIC_URL`
as a prefix).